requirement.txt
adbutils==0.8.2
aircv==1.4.6
airtest==1.1.4
allure-pytest==2.8.6
allure-python-commons==2.8.6
apipkg==1.5
apkutils2==1.0.0
Appium-Python-Client==1.0.1
attrs==19.3.0
cached-property==1.5.1
certifi==2020.6.20
chardet==3.0.4
cigam==0.0.3
decorator==4.4.2
Deprecated==1.2.10
deprecation==2.1.0
execnet==1.7.1
Faker==4.1.1
grpcio==1.30.0
idna==2.10
importlib-metadata==1.7.0
Jinja2==2.11.2
logzero==1.5.0
lxml==4.5.2
MarkupSafe==1.1.1
more-itertools==8.4.0
mss==4.0.3
numpy==1.19.1
opencv-contrib-python==3.4.2.17
opencv-python==4.3.0.36
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
progress==1.5
py==1.9.0
pyelftools==0.26
pyparsing==2.4.7
pytest==5.3.5
pytest-forked==1.2.0
pytest-html==2.0.0
pytest-metadata==1.10.0
pytest-ordering==0.6
pytest-parallel==0.1.0
pytest-rerunfailures==8.0
pytest-xdist==1.31.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-xlib==0.27
pywinauto==0.6.3
PyYAML==5.3
requests==2.24.0
retry==0.9.2
selenium==3.141.0
six==1.15.0
tblib==1.6.0
text-unidecode==1.3
uiautomator2==2.11.2
urllib3==1.25.9
wcwidth==0.2.5
whichcraft==0.6.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.1.0